pyproject.toml
[project]
name = "aurelia"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord>=2.3.2",
    "discord-py[voice]>=2.4.0",
    "pynacl>=1.5.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]

[project.scripts]
aurelia = "aurelia:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
